[Merge] new material and textures
Now a TriMesh may contains many submeshes (as .obj file may have mutiple objects).
Each submesh may have bound Materials.

Current Material is a combination of shaders and textures, as different materials requires different shaders.
This is a tough integration but it is more difficult to expose API for a mutilple object .obj and .mtl file.